Use Mutex instead of atomic integers to have floats in counter and gauge

+++ src/protocol.rs
@@ -260,10 +260,10 @@ pub enum RillData {
         message: String,
     },
     CounterRecord {
-        value: u64,
+        value: f64,
     },
     GaugeValue {
-        value: u64,
+        value: f64,
     },
 }
